@@ -73,72 +75,74 @@
     "github-username-regex": "^1.0.0",
     "hsts": "2.2.0",
     "language-map": "1.4.0",
-    "lodash": "4.17.19",
-    "luxon": "1.24.1",
+    "lodash": "4.17.20",
+    "luxon": "1.25.0",
     "memory-cache": "0.2.0",
-    "moment": "2.27.0",
-    "moment-timezone": "0.5.31",
+    "moment": "2.29.1",
+    "moment-timezone": "0.5.32",
     "morgan": "1.10.0",
     "mz": "2.7.0",
-    "node-jose": "1.1.4",
-    "nodemailer": "6.4.11",
+    "node-jose": "2.0.0",
+    "nodemailer": "6.4.17",
     "oauth": "0.9.x",
     "object-path": "0.11.5",
     "octicons": "5.0.1",
     "only": "0.0.2",
     "package-lock.json": "1.0.0",
-    "painless-config-resolver": "4.0.4",
+    "painless-config-resolver": "4.0.5",
     "passport": "0.4.1",
-    "passport-azure-ad": "4.2.1",
+    "passport-azure-ad": "4.3.0",
     "passport-github": "1.1.0",
     "passport-strategy": "1.x.x",
-    "pg": "8.3.0",
+    "pg": "8.5.1",
     "pg-escape": "0.2.0",
-    "pug": "2.0.4",
+    "pug": "3.0.0",
     "recursive-readdir": "2.2.2",
     "redis": "3.0.2",
-    "redis-mock": "0.50.0",
+    "redis-mock": "0.56.3",
     "request": "2.88.2",
     "request-promise-native": "1.0.9",
-    "requestretry": "4.1.1",
+    "requestretry": "4.1.2",
     "secure-compare": "3.0.1",
-    "semver": "7.3.2",
+    "semver": "7.3.4",
     "serve-favicon": "2.5.0",
+    "simple-oauth2": "4.2.0",
     "throat": "5.0.0",
     "tmp": "0.2.1",
     "tmp-promise": "3.0.2",
     "unzipper": "^0.10.11",
-    "uuid": "8.3.0",
-    "validator": "13.1.1"
+    "uuid": "8.3.2",
+    "validator": "13.5.2"
   },
   "devDependencies": {
-    "@types/azure-sb": "0.0.37",
-    "@types/cheerio": "0.22.21",
-    "@types/connect-redis": "0.0.14",
-    "@types/core-js": "2.5.3",
+    "@types/azure-sb": "0.0.38",
+    "@types/cheerio": "0.22.23",
+    "@types/connect-redis": "0.0.16",
+    "@types/core-js": "2.5.4",
     "@types/debug": "4.1.5",
-    "@types/express": "^4.17.7",
-    "@types/express-session": "1.17.0",
-    "@types/lodash": "4.14.159",
-    "@types/luxon": "1.24.3",
+    "@types/express": "4.17.9",
+    "@types/express-session": "1.17.3",
+    "@types/lodash": "4.14.167",
+    "@types/luxon": "1.25.0",
     "@types/memory-cache": "0.2.1",
-    "@types/mocha": "8.0.1",
-    "@types/node": "14.0.27",
-    "@types/node-jose": "1.1.4",
-    "@types/passport": "^1.0.4",
+    "@types/mocha": "8.2.0",
+    "@types/node": "14.14.20",
+    "@types/node-jose": "1.1.5",
+    "@types/passport": "1.0.5",
     "@types/passport-azure-ad": "^4.0.7",
     "@types/passport-github": "^1.1.5",
     "@types/passport-strategy": "0.2.35",
-    "@types/pg": "7.14.4",
+    "@types/pg": "7.14.7",
     "@types/recursive-readdir": "2.2.0",
-    "@types/redis": "2.8.25",
+    "@types/redis": "2.8.28",
     "@types/request": "2.48.5",
     "@types/request-promise-native": "1.0.17",
+    "@types/simple-oauth2": "4.1.0",
     "@types/unzipper": "^0.10.3",
-    "@types/uuid": "8.0.1",
-    "@types/validator": "13.1.0",
+    "@types/uuid": "8.3.0",
+    "@types/validator": "13.1.2",
     "chai": "4.2.0",
-    "mocha": "8.1.1",
-    "typescript": "3.9.7"
+    "mocha": "8.2.1",
+    "typescript": "4.1.3"
   }
 }
