@@ -5,7 +5,7 @@
 
 import { PostgresLinkProvider } from './postgresLinkProvider';
 
-module.exports = function createProvider(providers, config) {
+export default function createPostgresProvider(providers, config) {
   if (!providers.postgresPool) {
     throw new Error('A Postgres pool must be initialized and available at providers.postgresPool to use the PostgresLinkProvider');
   }
@@ -24,4 +24,4 @@ module.exports = function createProvider(providers, config) {
   }
 
   return new PostgresLinkProvider(providers, options);
-};
+}
