@@ -3,7 +3,7 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 //
 
-const querystring = require('querystring');
+import querystring from 'querystring';
 
 import { redirectToReferrer, storeReferrer } from '../utils';
 import { getGithubAppConfigurationOptions } from './passport/githubStrategy';
@@ -32,8 +32,25 @@ function newSessionAfterAuthentication(req, res, next) {
 }
 
 module.exports = function configurePassport(app, passport, initialConfig) {
-  app.get('/signin', function (req, res) {
-    storeReferrer(req, res, '/auth/azure', 'signin page hit, need to go authenticate');
+  app.get('/signin', function (req, res, next) {
+    if (req.isAuthenticated()) {
+      const username = req.user?.azure?.username;
+      if (username) {
+        // Do not sign in yet again
+        const nextDestination = req.headers?.referer || '/';
+        return res.redirect(nextDestination);
+      }
+    }
+    const currentlyStoredSessionReferer = req.session?.referer || undefined;
+    req.session.regenerate(function (err) {
+      if (err) {
+        return next(err);
+      }
+      if (currentlyStoredSessionReferer && req.session) {
+        req.session.referer = currentlyStoredSessionReferer;
+      }
+      storeReferrer(req, res, '/auth/azure', 'signin page hit, need to go authenticate');
+    });
   });
 
   // ----------------------------------------------------------------------------
@@ -160,7 +177,7 @@ module.exports = function configurePassport(app, passport, initialConfig) {
     });
   }
 
-  app.get('/signout', function (req, res) {
+  function signoutPage(req, res) {
     var config = req.app.settings.runtimeConfig;
     req.logout();
     if (req.session) {
@@ -178,7 +195,10 @@ module.exports = function configurePassport(app, passport, initialConfig) {
         config: initialConfig.obfuscatedConfig,
       });
     }
-  });
+  };
+
+  app.get('/signout',signoutPage);
+  app.get('/signout/goodbye',signoutPage);
 
   app.get('/signout/github', processSignout.bind(null, 'github', 'github,githubIncreasedScope'));
 
@@ -231,15 +251,17 @@ module.exports = function configurePassport(app, passport, initialConfig) {
       });
     }
     const messageError: IReposError = new Error(
-      isAuthenticated ? 'Authentication initially failed, but you are good to go now.' : 'Authentication failed, possibly due to a state problem. This can happen when certain tools or apps launch URLs. Try signing in again now.');
+      isAuthenticated ? 'Authentication initially failed, but you are good to go now.' : 'Authentication failed, possibly due to SameSite cookie issues with newer browsers.');
     messageError.skipLog = true;
+    messageError.status = 400;
     if (isAuthenticated) {
-      messageError.skipOops = true;
-      messageError.detailed = 'Unfortunately we were not able to take you to the URL that you clicked on. If you go to that URL now, your request should work!';
-      messageError.fancyLink = {
-        link: '/',
-        title: 'Go to the site homepage',
-      };
+      return res.redirect('/');
+      // messageError.skipOops = true;
+      // messageError.detailed = 'Unfortunately we were not able to take you to the URL that you clicked on. If you go to that URL now, your request should work!';
+      // messageError.fancyLink = {
+      //   link: '/',
+      //   title: 'Go to the site homepage',
+      // };
     } else {
       messageError.fancyLink = {
         link: '/auth/azure',
