@@ -12,6 +12,7 @@ import { wrapError } from '../../../../utils';
 import { Team } from '../../../../business/team';
 import { PermissionWorkflowEngine } from '../approvals';
 import RenderHtmlMail from '../../../../lib/emailRender';
+import { IndividualContext } from '../../../../user';
 
 interface ILocalRequest extends ReposAppRequest {
   team2?: any;
@@ -59,130 +60,142 @@ router.get('/setNote/:action', function (req: ILocalRequest, res) {
   });
 });
 
-router.post('/', asyncHandler(async function (req: ILocalRequest, res, next) {
+router.post('/', asyncHandler(async (req: ILocalRequest, res, next) => {
+  const providers = req.app.settings.providers as IProviders;
+  const { individualContext } = req;
   const engine = req.approvalEngine as PermissionWorkflowEngine;
-  const approvalRequest = req.approvalEngine.request;
-  const requestid = engine.id;
-  const {
-    approvalProvider: teamJoinApprovalProvider,
-    config,
-    mailAddressProvider,
-    mailProvider,
-  } = req.app.settings.providers as IProviders;
-  if (!req.body.text && req.body.deny) {
-    return res.redirect(req.teamUrl + 'approvals/' + requestid + '/setNote/deny');
-  }
+  let message = req.body.text as string;
+  const teamBaseUrl = req.teamUrl as string;
+  let decision: TeamApprovalDecision = null;
   if (req.body.reopen) {
-    approvalRequest.active = true;
-    teamJoinApprovalProvider.updateTeamApprovalEntity(approvalRequest).then(ok => {
-      req.individualContext.webContext.saveUserAlert('Request re-opened.', engine.typeName, UserAlertType.Success);
-      res.redirect(req.teamUrl + 'approvals/' + requestid);
+    decision = TeamApprovalDecision.Reopen;
+  } else if (req.body.approve || req.body.approveWithComment) {
+    decision = TeamApprovalDecision.Approve;
+  } else if (req.body.deny) {
+    decision = TeamApprovalDecision.Deny;
+  }
+  if (!decision) {
+    throw new Error('No valid decision');
+  }
+  const outcome = await postActionDecision(providers, individualContext, engine, teamBaseUrl, decision, message);
+  if (outcome.message) {
+    req.individualContext.webContext.saveUserAlert(outcome.message, engine.typeName, UserAlertType.Success);
+  }
+  if (outcome.error) {
+    req.insights.trackException({
+      exception: outcome.error,
     });
+    return next(outcome.error);
   }
-  if (!req.body.text && req.body.approveWithComment) {
-    return res.redirect(req.teamUrl + 'approvals/' + requestid + '/setNote/approveWithComment');
+  return res.redirect(outcome.redirect || teamBaseUrl);
+}));
+
+export enum TeamApprovalDecision {
+  Approve = 'Approved',
+  Deny = 'Denied',
+  Reopen = 'Reopened',
+}
+
+export interface IPostActionDecisionOutcome {
+  redirect?: string;
+  message?: string;
+  error?: any;
+}
+
+function performApprovalWithEngine(engine: PermissionWorkflowEngine): Promise<void> {
+  return new Promise((resolve, reject) => {
+    engine.performApprovalOperation((error: Error) => {
+      return error ? reject(error) : resolve();
+    });
+  });
+}
+
+export async function postActionDecision(providers: IProviders, individualContext: IndividualContext, engine: PermissionWorkflowEngine, teamBaseUrl: string, decision: TeamApprovalDecision, messageToRequestor: string): Promise<IPostActionDecisionOutcome> {
+  if (!individualContext || !individualContext.getGitHubIdentity().username) {
+    return { error: new Error('No individual context') };
   }
-  const repoApprovalTypesValues = config.github.approvalTypes.repo;
-  if (repoApprovalTypesValues.length === 0) {
-    return next(new Error('No repo approval providers configured.'));
+  const approvalRequest = engine.request;
+  const requestid = engine.id;
+  const { approvalProvider: teamJoinApprovalProvider, config, mailAddressProvider, mailProvider, insights } = providers;
+  if (decision === TeamApprovalDecision.Reopen) {
+    approvalRequest.active = true;
+    try {
+      await teamJoinApprovalProvider.updateTeamApprovalEntity(approvalRequest);
+      return {
+        message: 'Request reopened',
+        redirect: `${teamBaseUrl}approvals/${requestid}`,
+      };
+    } catch (error) {
+      return { error };
+    }
   }
-  const repoApprovalTypes = new Set(repoApprovalTypesValues);
-  const mailProviderInUse = repoApprovalTypes.has('mail');
-  if (!mailProviderInUse) {
-    return next(new Error('No configured approval providers configured.'));
-  }
-  if (!mailProvider) {
-    return next(new Error('A mail provider has been requested but a provider instance could not be found.'));
-  }
-  // Approval workflow note: although the configuration may specify just a mail
-  // provider today, there may actually be an issue that was opened at the time
-  // of the request. So we will attempt to close any issues if the request has
-  // an issue ID.
-  const action = req.body.approveWithComment || req.body.approve ? 'approve' : 'deny';
-  const bodyText = req.body.text;
-  const username = req.individualContext.getGitHubIdentity().username;
-  let friendlyErrorMessage = 'Whoa? What happened?';
+  const action = decision === TeamApprovalDecision.Approve ? 'approve' : 'deny';
+  const username = individualContext.getGitHubIdentity().username;
+  let userMailAddress: string = null;
+  const decisionMessage = messageToRequestor || decision;
   const pendingRequest = engine.request;
-  const upn = pendingRequest.corporateUsername;
-  let userMailAddress = null;
   try {
+    const upn = pendingRequest.corporateUsername;
     userMailAddress = await mailAddressProvider.getAddressFromUpn(upn);
     pendingRequest.decision = action;
     pendingRequest.active = false;
     pendingRequest.decisionTime = new Date();
     pendingRequest.decisionThirdPartyUsername = username;
-    pendingRequest.decisionThirdPartyId = req.individualContext.getGitHubIdentity().id;
-    pendingRequest.decisionMessage = bodyText;
-    pendingRequest.decisionCorporateUsername = req.individualContext.corporateIdentity.username;
-    pendingRequest.decisionCorporateId = req.individualContext.corporateIdentity.id;
-    friendlyErrorMessage = 'The approval request information could not be updated, indicating a data store problem potentially. The decision may not have been recorded.';
+    pendingRequest.decisionThirdPartyId = individualContext.getGitHubIdentity().id;
+    pendingRequest.decisionMessage = decisionMessage;
+    pendingRequest.decisionCorporateUsername = individualContext.corporateIdentity.username;
+    pendingRequest.decisionCorporateId = individualContext.corporateIdentity.id;
     await teamJoinApprovalProvider.updateTeamApprovalEntity(pendingRequest);
-    if (action == 'approve') {
-      await new Promise((resolve, reject) => {
-        engine.performApprovalOperation((error: Error) => {
-          return error ? reject(error) : resolve();
-        });
-      });
+    if (decision == TeamApprovalDecision.Approve) {
+      await performApprovalWithEngine(engine);
     }
-    friendlyErrorMessage = null;
   } catch (error) {
-    if (friendlyErrorMessage) {
-      error = wrapError(error, friendlyErrorMessage);
-    }
-    return next(error);
+    return { error };
   }
-  let secondaryErrors = false;
-  req.individualContext.webContext.saveUserAlert('Thanks for your ' + action.toUpperCase() + ' decision.', engine.typeName, UserAlertType.Success);
-  if (mailProviderInUse) {
-    const wasApproved = action === 'approve';
+  const message = `Thanks for your ${action.toUpperCase()} decision`;
+  const approvalMail = individualContext.link.corporateMailAddress || individualContext.link.corporateUsername;
+  if (mailProvider) {
+    const wasApproved = decision === TeamApprovalDecision.Approve;
     const contentOptions = {
-      correlationId: req.correlationId,
+      correlationId: individualContext.webContext?.correlationId,
       pendingRequest,
       version: config.logging.version,
       results: [], // no longer a used field, used to be called 'output'
       wasApproved,
       decisionBy: username,
-      decisionNote: bodyText,
-      decisionEmail: req.individualContext.corporateIdentity.username,
+      decisionNote: decisionMessage,
+      decisionEmail: approvalMail,
       reason: (`You are receiving this e-mail because of a request that you created, and a decision has been made.
-                This mail was sent to: ${pendingRequest.corporateUsername}`),
+                This mail was sent to: ${userMailAddress}`),
       headline: engine.getDecisionEmailHeadline(wasApproved),
       notification: wasApproved ? 'information' : 'warning',
-      service: (config.brand?.companyName || 'Microsoft') + ' GitHub',
+      service: (config.brand?.companyName || 'Corporate') + ' GitHub',
     };
     if (!engine.getDecisionEmailViewName || !engine.getDecisionEmailSubject) {
-      return req.insights.trackException({
-        exception: new Error('No getDecisionEmailViewName available with the engine.'),
-        properties: Object.assign({ eventName: 'ReposRequestDecisionMailRenderFailure' }, contentOptions),
-      });
+      return { message, redirect: teamBaseUrl };
     }
     const getDecisionEmailViewName = engine.getDecisionEmailViewName();
     let content = null;
     try {
-      content = await RenderHtmlMail(req.app.settings.runtimeConfig.typescript.appDirectory, getDecisionEmailViewName, contentOptions);
+      content = await RenderHtmlMail(config.typescript.appDirectory, getDecisionEmailViewName, contentOptions);
     } catch (renderError) {
-      req.insights.trackException({
-        exception: renderError,
-        properties: Object.assign({ eventName: 'ReposRequestDecisionMailRenderFailure' }, contentOptions),
-      });
     }
     if (content) {
       const mail = {
         to: [userMailAddress],
         subject: engine.getDecisionEmailSubject(wasApproved, pendingRequest),
         content,
-        correlationId: req.correlationId,
-        category: ['decision', 'repos'],
+        correlationId: individualContext.webContext?.correlationId,
       };
       try {
         const mailResult = await mailProvider.sendMail(mail);
-        req.insights.trackEvent({
+        insights?.trackEvent({
           name: 'ReposRequestDecisionMailSuccess', properties: Object.assign({
             receipt: mailResult,
           }, contentOptions)
         });
       } catch (mailError) {
-        req.insights.trackException({
+        insights?.trackException({
           exception: mailError, properties: Object.assign({
             eventName: 'ReposRequestDecisionMailFailure',
           }, contentOptions)
@@ -190,7 +203,7 @@ router.post('/', asyncHandler(async function (req: ILocalRequest, res, next) {
       }
     }
   }
-  return res.redirect(req.teamUrl);
-}));
+  return { message, redirect: teamBaseUrl };
+}
 
-module.exports = router;
+export default router;
