@@ -8,11 +8,25 @@ import asyncHandler from 'express-async-handler';
 
 import { jsonError } from '../../middleware/jsonError';
 import { apiContextMiddleware } from '../../middleware/business/setContext';
-import { setIdentity } from '../../middleware/business/authentication';
+import { requireAccessTokenClient, setIdentity } from '../../middleware/business/authentication';
 import { AddLinkToRequest } from '../../middleware/links';
 import { ReposAppRequest } from '../../transitional';
 
+import RouteEmberClientNewRepo from './newRepo';
+
 import ReleaseApprovalsRoute from './releaseApprovals';
+
+import RouteServiceTree from './internal/serviceTree';
+import RouteDirectory from './internal/directory';
+import RouteOrganizations from './internal/organizations';
+import RouteContext from './internal/context';
+import RouteLinking from './internal/linking';
+import RouteSession from './internal/session';
+import RouteBanner from './internal/banner';
+import RouteCrossOrganizationPeople from './internal/people';
+import RouteCrossOrganizationRepos from './internal/repos';
+import RouteCrossOrganizationTeams from './internal/teams';
+import RouteCorporateRepoMetadata from './internal/corporateRepoMetadata';
 
 const router = express.Router();
 
@@ -22,12 +36,26 @@ router.use((req: ReposAppRequest, res, next) => {
   }
   return next(jsonError('The current session is not authenticated', 401));
 });
-
+router.use(asyncHandler(requireAccessTokenClient));
 router.use(apiContextMiddleware);
 router.use(setIdentity);
 router.use(asyncHandler(AddLinkToRequest));
 
-router.use('/newRepo', require('./newRepo'));
+// --- new React client ---
+router.use('/serviceTree', RouteServiceTree);
+router.use('/banner', RouteBanner);
+router.use('/directory', RouteDirectory);
+router.use('/orgs', RouteOrganizations);
+router.use('/context', RouteContext);
+router.use('/link', RouteLinking);
+router.use('/signout', RouteSession);
+router.use('/corporateRepoMetadata', RouteCorporateRepoMetadata);
+router.use('/people', RouteCrossOrganizationPeople);
+router.use('/repos', RouteCrossOrganizationRepos);
+router.use('/teams', RouteCrossOrganizationTeams);
+// --- end of new React client work ---
+
+router.use('/newRepo', RouteEmberClientNewRepo);
 router.use('/releaseApprovals', ReleaseApprovalsRoute);
 
 router.use((req, res, next) => {
