@@ -298,7 +298,7 @@ export class Operations {
     // const hasConfiguredOrganizations = this.config.github.organizations && this.config.github.organizations.length;
     const organizationSettingsProvider = this.providers.organizationSettingsProvider;
     if (hasModernGitHubApps && organizationSettingsProvider) {
-      const dynamicOrganizations = (await organizationSettingsProvider.queryAllOrganizations()).filter(dynamicOrg => dynamicOrg.active === true);
+      const dynamicOrganizations = (await organizationSettingsProvider.queryAllOrganizations()).filter(dynamicOrg => dynamicOrg.active === true && !dynamicOrg.hasFeature('ignore'));
       this._dynamicOrganizationSettings = dynamicOrganizations;
       this._dynamicOrganizationIds = new Set(dynamicOrganizations.map(org => asNumber(org.organizationId)));
     }
@@ -415,7 +415,7 @@ export class Operations {
   }
 
   private async getAppAuthorizationHeader(tokenManager: GitHubTokenManager, appId: number): Promise<string> {
-    const jwt = tokenManager.getAppById(appId).getSignedJsonWebToken();
+    const jwt = await tokenManager.getAppById(appId).getAppAuthenticationToken();
     const value = `bearer ${jwt}`;
     return value;
   }
@@ -575,7 +575,10 @@ export class Operations {
   }
 
   validateGitHubLogin(username: string) {
-    if (!githubUsernameRegex.test(username)) {
+    // There are some legitimate usernames at GitHub that have a dash
+    // in them. While GitHub no longer allows this for new accounts,
+    // they are grandfathered in.
+    if (!githubUsernameRegex.test(username) && !username.endsWith('-')) {
       throw new Error('Invalid GitHub username format');
     }
     return username;
@@ -679,6 +682,11 @@ export class Operations {
         mailAddress = corporateAccount.mail || link.serviceAccountMail;
         link.corporateDisplayName = corporateAccount.displayName;
         link.corporateUsername = corporateAccount.userPrincipalName;
+        link.corporateMailAddress = corporateAccount.mail;
+        // NOTE: this is strongly typed to the AAD graph info response right now instead of more generic
+        if (corporateAccount.mailNickname) {
+          link.corporateAlias = corporateAccount.mailNickname.toLowerCase();
+        }
         // Validate that the corporate account can be linked
         if (corporateInfo.type === SupportedLinkType.ServiceAccount) {
           if (!link.serviceAccountMail) {
@@ -1088,14 +1096,23 @@ export class Operations {
 
   getOrganization(name: string): Organization {
     if (!name) {
-      throw new Error('getOrganization: name required');
+      throw CreateError.ParameterRequired('name');
     }
     const lc = name.toLowerCase();
     const organization = this.organizations.get(lc);
     if (!organization) {
-      throw new Error(`Could not find configuration for the "${name}" organization.`);
+      throw CreateError.NotFound(`Could not find configuration for the "${name}" organization.`);
     }
     return organization;
+  }
+
+  isOrganizationManagedById(organizationId: number): boolean {
+    try {
+      this.getOrganizationById(organizationId);
+      return true;
+    } catch (notConfigured) {
+      return false;
+    }
   }
 
   getOrganizationById(organizationId: number): Organization {
@@ -1108,7 +1125,7 @@ export class Operations {
     }
     const org = this._organizationIds.get(organizationId);
     if (!org) {
-      throw new Error(`getOrganizationById: no configured ID for an organization with ID ${organizationId}`);
+      throw CreateError.NotFound(`getOrganizationById: no configured ID for an organization with ID ${organizationId}`);
     }
     return org;
   }
@@ -1417,7 +1434,7 @@ export class Operations {
     options = options || {};
     const operations = this;
     if (!username) {
-      throw new Error('Must provide a GitHub username to retrieve account information.');
+      throw CreateError.ParameterRequired('username');
     }
     const parameters = {
       username: username,
