@@ -14,7 +14,13 @@ export default async function initializeAlternateApps(config, app: IReposApplica
   const appPath = path.resolve(path.join(__dirname, '..', appName, '/'));
   debug(`Alternate app requested: name=${appName}, path=${appPath}`);
   try {
-    const setupApp = require(appPath);
+    let setupApp = require(appPath);
+    // support modern imports
+    if (typeof(setupApp) !== 'function' && setupApp.default) {
+      setupApp = setupApp.default
+    } else if (typeof(setupApp) !== 'function') {
+      throw new Error(`Could not prepare default import for alternate app name=${appName}, path=${appPath}`);
+    }
     return await setupApp(config, app);
   } catch (loadAlternateAppError) {
     console.log(`Loading error: ${loadAlternateAppError}`);
