@@ -7,7 +7,8 @@
 
 import _ from 'lodash';
 const debug = require('debug')('user');
-import { ReposAppRequest } from '../../transitional';
+
+import { IAppSession, IProviders, ReposAppRequest } from '../../transitional';
 import { ICorporateIdentity, IGitHubIdentity, IndividualContext, GitHubIdentitySource } from '../../user';
 import { storeOriginalUrlAsReferrer } from '../../utils';
 
@@ -20,6 +21,45 @@ export function requireAuthenticatedUserOrSignInExcluding(exclusionPaths: string
     }
   }
   return requireAuthenticatedUserOrSignIn(req, res, next);
+}
+
+export async function requireAccessTokenClient(req: ReposAppRequest, res, next) {
+  if (req.oauthAccessToken) {
+    return next();
+  }
+  // This code currently assumes you're using AAD.
+  const { authorizationCodeClient } = req.app.settings.providers as IProviders;
+  if (!req.user.azure) {
+    console.warn('Not an Azure authenticated user');
+    return signoutThenSignIn(req, res);
+  }
+  // Build an OAuth Access Token instance for the request, refreshing as needed
+  const { oauthToken } = req.user.azure;
+  if (authorizationCodeClient && oauthToken) {     
+    const hydratedToken = JSON.parse(oauthToken);
+    let oauthTokenInstance = authorizationCodeClient.createToken(hydratedToken);
+    if (oauthTokenInstance.expired()) {
+      oauthTokenInstance = await oauthTokenInstance.refresh();
+      const session = req.session as IAppSession;
+      session.passport.user.azure.oauthToken = JSON.stringify(oauthTokenInstance.token);
+    }
+    req.oauthAccessToken = oauthTokenInstance;
+  } else {
+    // this is only used during the transition to storing this kind of data
+    // console.warn('The user did not have an oauthToken available');
+    // return signoutThenSignIn(req, res);
+  }
+  return next();
+}
+
+function signoutThenSignIn(req, res) {
+  req.logout();
+  return redirectToSignIn(req, res);
+}
+
+function redirectToSignIn(req, res) {
+  const config = req.app.settings.runtimeConfig;
+  storeOriginalUrlAsReferrer(req, res, config.authentication.scheme === 'github' ? '/auth/github' : '/auth/azure', 'user is not authenticated and needs to authenticate');
 }
 
 export function requireAuthenticatedUserOrSignIn(req: ReposAppRequest, res, next) {
@@ -36,7 +76,7 @@ export function requireAuthenticatedUserOrSignIn(req: ReposAppRequest, res, next
     }
     return next();
   }
-  storeOriginalUrlAsReferrer(req, res, config.authentication.scheme === 'github' ? '/auth/github' : '/auth/azure', 'user is not authenticated and needs to authenticate');
+  return redirectToSignIn(req, res);
 }
 
 export function setIdentity(req: ReposAppRequest, res, next) {
