@@ -3,6 +3,10 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 //
 
+const debug = require('debug')('startup');
+
+import { AuthorizationCode } from 'simple-oauth2';
+
 import { IProviders } from '../../transitional';
 
 import { OIDCStrategy } from 'passport-azure-ad';
@@ -16,19 +20,24 @@ interface IAADUser {
   displayName: string;
   oid: string;
   username: string;
+  oauthToken?: string;
 }
 
-async function login(app, config, iss, sub, profile): Promise<IPassportUserWithAAD> {
+async function login(app, config, client: AuthorizationCode, iss, sub, profile, accessToken: string, refreshToken: string, params): Promise<IPassportUserWithAAD> {
   const { graphProvider } = app.settings.providers as IProviders;
+  const oauthToken = JSON.stringify(params);
   if (config && config.impersonation && config.impersonation.corporateId) {
+    // While impersonation for the site interface is possible, the graph API token,
+    // rarely used in this app, will use the actual AAD access tokens still.
     const impersonationCorporateId = config.impersonation.corporateId;
     const impersonationResult = await graphProvider.getUserById(impersonationCorporateId);
-    console.warn(`IMPERSONATION: id=${impersonationResult.id} upn=${impersonationResult.userPrincipalName} name=${impersonationResult.displayName}`);
+    console.warn(`IMPERSONATION: id=${impersonationResult.id} upn=${impersonationResult.userPrincipalName} name=${impersonationResult.displayName} graphIsNotImpersonatedAs=${profile.upn}`);
     return {
       azure: {
         displayName: impersonationResult.displayName,
         oid: impersonationResult.id,
         username: impersonationResult.userPrincipalName,
+        oauthToken,
       },
     };
   }
@@ -43,12 +52,13 @@ async function login(app, config, iss, sub, profile): Promise<IPassportUserWithA
       displayName: profile.displayName,
       oid: profile.oid,
       username: profile.upn,
+      oauthToken,
     },
   };
 }
 
-function activeDirectorySubset(app, config, iss, sub, profile, done) {
-  login(app, config, iss, sub, profile).then(profile => {
+function activeDirectorySubset(app, config, client: AuthorizationCode, iss, sub, profile, accessToken: string, refreshToken: string, params, done) {
+  login(app, config, client, iss, sub, profile, accessToken, refreshToken, params).then(profile => {
     return done(null, profile);
   }).catch(error => {
     return done(error);
@@ -57,19 +67,42 @@ function activeDirectorySubset(app, config, iss, sub, profile, done) {
 
 export default function createAADStrategy(app, config) {
   const { redirectUrl, tenantId, clientId, clientSecret } = config.activeDirectory;
-  let aadStrategy = new OIDCStrategy({
+  const providers = app.settings.providers as IProviders;
+  const aadAuthority = `https://login.microsoftonline.com/${tenantId}/`;
+  const aadMetadata =  'v2.0/.well-known/openid-configuration'; // used to use: .well-known/openid-configuration
+  const identityMetadata = `${aadAuthority}${aadMetadata}`;
+  const originalIdentityMetadata = `${aadAuthority}.well-known/openid-configuration`;
+  const authorizePath = 'oauth2/v2.0/authorize';
+  const tokenPath = 'oauth2/v2.0/token';
+  const aadScopes = 'profile openid user.read';
+  const oauth2Client = new AuthorizationCode({
+    client: {
+      id: clientId,
+      secret: clientSecret,
+    },
+    auth: {
+      tokenHost: aadAuthority,
+      tokenPath,
+      authorizePath,
+    },
+  });
+  debug(`AAD app clientId=${clientId}, redirectUrl=${redirectUrl}`);
+  providers.authorizationCodeClient = oauth2Client;
+  const aadStrategy = new OIDCStrategy({
     redirectUrl: redirectUrl || `${config.webServer.baseUrl}/auth/azure/callback`,
     allowHttpForRedirectUrl: config.containers.docker || config.webServer.allowHttp,
     // @ts-ignore
     realm: tenantId,
     clientID: clientId,
-    clientSecret: clientSecret,
-    identityMetadata: `https://login.microsoftonline.com/${tenantId}/.well-known/openid-configuration`,
+    clientSecret,
+    identityMetadata: originalIdentityMetadata,
     responseType: 'id_token code',
     responseMode: 'form_post',
+    scope: aadScopes.split(' '),
+    // cookieSameSite: true, // ???
     // oidcIssuer: config.activeDirectory.issuer,
     // validateIssuer: true,
-  }, activeDirectorySubset.bind(null, app, config));
+  }, activeDirectorySubset.bind(null, app, config, oauth2Client));
   // Patching the AAD strategy to intercept a specific state failure message and instead
   // of providing a generic failure message, redirecting (HTTP GET) to the callback page
   // where we can offer a more useful message
