@@ -7,7 +7,7 @@ import express from 'express';
 import asyncHandler from 'express-async-handler';
 const router = express.Router();
 
-import { ReposAppRequest, IProviders, CreateError } from '../transitional';
+import { ReposAppRequest, IProviders, CreateError, hasStaticReactClientApp } from '../transitional';
 import { IndividualContext } from '../user';
 import { storeOriginalUrlAsVariable } from '../utils';
 import { AuthorizeOnlyCorporateAdministrators } from '../middleware/business/corporateAdministrators';
@@ -17,6 +17,7 @@ import { Organization } from '../business/organization';
 import { Repository } from '../business/repository';
 
 import orgsRoute from './orgs';
+import { injectReactClient } from '../microsoft/preview';
 
 const orgAdmin = require('./orgAdmin');
 const peopleRoute = require('./people');
@@ -24,7 +25,9 @@ const setupRoute = require('./administration');
 const reposRoute = require('./repos');
 const teamsRoute = require('./teams');
 const undoRoute = require('./undo');
-const contributionsRoute = require('./contributions');
+
+const hasReactApp = hasStaticReactClientApp();
+const reactRoute = hasReactApp ? injectReactClient() : undefined;
 
 //-----------------------------------------------------------------------------
 //-----------------------------------------------------------------------------
@@ -32,27 +35,32 @@ const contributionsRoute = require('./contributions');
 // Below this next call, all routes will require an active link to exist for
 // the authenticated GitHub user.
 //-----------------------------------------------------------------------------
+// * only for the traditional app. The React app does not require a link to browse orgs.
 //-----------------------------------------------------------------------------
-router.use(function (req: ReposAppRequest, res, next) {
-  const individualContext = req.individualContext as IndividualContext;
-  const link = individualContext.link;
-  if (link && link.thirdPartyId) {
-    return next();
-  }
-  storeOriginalUrlAsVariable(req, res, 'beforeLinkReferrer', '/', 'no linked github username');
-});
+if (!hasReactApp) {
+  router.use(function (req: ReposAppRequest, res, next) {
+    const individualContext = req.individualContext as IndividualContext;
+    const link = individualContext.link;
+    if (link && link.thirdPartyId) {
+      return next();
+    }
+    storeOriginalUrlAsVariable(req, res, 'beforeLinkReferrer', '/', 'no linked github username');
+  });
+}
 // end security route
 //-----------------------------------------------------------------------------
 //-----------------------------------------------------------------------------
 
 router.use('/unlink', unlinkRoute);
 
-router.use('/teams', teamsRoute);
-router.use('/organization', orgAdmin);
-router.use('/people', peopleRoute);
-router.use('/repos', reposRoute);
+router.use('/organization', orgAdmin); // admin UI, not in React
+
+router.use('/teams', reactRoute || teamsRoute);
+router.use('/people', reactRoute || peopleRoute);
+router.use('/repos', reactRoute ||reposRoute);
+
+// Routes not yet available in the client
 router.use('/undo', undoRoute);
-router.use('/contributions', contributionsRoute);
 router.use('/administration', AuthorizeOnlyCorporateAdministrators, setupRoute);
 
 router.use('/https?*github.com/:org/:repo', asyncHandler(async (req, res, next) => {
@@ -73,6 +81,9 @@ router.use('/https?*github.com/:org/:repo', asyncHandler(async (req, res, next) 
     }
     catch (error) {
       return next(CreateError.NotFound(`The repository ${org}/${repo} no longer exists.`));
+    }
+    if (hasReactApp) {
+      return res.redirect(`/orgs/${repository.organization.name}/repos/${repository.name}`);
     }
     return res.redirect(repository.baseUrl);
   }
