@@ -4,30 +4,60 @@
 //
 
 import express from 'express';
+import asyncHandler from 'express-async-handler';
 const router = express.Router();
 
-import { IReposRequestWithOrganization } from '../transitional';
+import querystring from 'querystring';
+
+import { hasStaticReactClientApp, IReposRequestWithOrganization } from '../transitional';
 import { wrapError } from '../utils';
 
 import orgRoute from './org/';
 
-router.use('/:orgName', function (req: IReposRequestWithOrganization, res, next) {
+import { injectReactClient, TryFallbackToBlob } from '../microsoft/preview';
+
+const hasReactApp = hasStaticReactClientApp();
+const reactRoute = hasReactApp ? injectReactClient() : undefined;
+
+if (hasReactApp) {
+  router.use('/orgs', reactRoute);
+  router.use('/orgs/:orgName', forwardToOrganizationRoutes);
+}
+
+router.use('/:orgName', asyncHandler(forwardToOrganizationRoutes));
+
+async function forwardToOrganizationRoutes (req: IReposRequestWithOrganization, res, next) {
   // This middleware contains both the original GitHub operations types
   // as well as the newer implementation. In time this will peel apart.
   const orgName = req.params.orgName;
   const operations = req.app.settings.operations;
   try {
-    req.organization = operations.getOrganization(orgName);
+    const organization = operations.getOrganization(orgName);
+    req.organization = organization;
+    if (hasReactApp && !req.path.includes('/byClient') /* special case */) {
+      const remainingPath = req.path;
+      let q = querystring.stringify(req.query as any);
+      q = q ? `?${q}` : '';
+      const reactClientPath = `/orgs/${organization.name}${remainingPath}${q}`;
+      console.log(`redirecting org route to react client route: ${reactClientPath}`);
+      return res.redirect(reactClientPath);
+    }
     return next();
   } catch (ex) {
     if (orgName.toLowerCase() == 'account') {
       return res.redirect('/');
     }
+    if (hasReactApp) {
+      const isComplete = await TryFallbackToBlob(req, res);
+      if (isComplete) {
+        return;
+      }
+    }
     const err = wrapError(null, 'Organization not found', true);
     err.status = 404;
     return next(err);
   }
-});
+}
 
 router.use('/:orgName', orgRoute);
 
